Change short type to ushort

+++ Shiptech.Domain/ValueObjects/AssortmentDictionaryLength.cs
@@ -1,7 +1,7 @@
 namespace Shiptech.Domain.ValueObjects;
 
-public record AssortmentDictionaryLength(double Value)
+public record AssortmentDictionaryLength(ushort Value)
 {
-    public static implicit operator double(AssortmentDictionaryLength length) => length.Value;
-    public static implicit operator AssortmentDictionaryLength(double length) => new(length);
+    public static implicit operator ushort(AssortmentDictionaryLength length) => length.Value;
+    public static implicit operator AssortmentDictionaryLength(ushort length) => new(length);
 };
